Moved the Container class.
