+++ src/Utils/DI/Container.php
@@ -12,7 +12,7 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon\Utils;
+namespace Jaxon\Utils\DI;
 
 use Lemon\Event\EventDispatcher;
 
@@ -74,8 +74,8 @@ class Container
     {
         $this->libContainer = new \Pimple\Container();
 
-        $sTranslationDir = realpath(__DIR__ . '/../../translations');
-        $sTemplateDir = realpath(__DIR__ . '/../../templates');
+        $sTranslationDir = realpath(__DIR__ . '/../../../translations');
+        $sTemplateDir = realpath(__DIR__ . '/../../../templates');
         $this->init($sTranslationDir, $sTemplateDir);
     }
